new state
New state called results.
So created results with null to return null. Then with response we update results with api data, then in search the results are rendered with a new component Results.
Use props to create a new component Meaning . Meaning when called uses props to loops through meaning, parts of speach and definition.